fix names and imports
