+++ backend/src/bounded-contexts/iam/iam/iam.module.ts
@@ -3,7 +3,6 @@ import { Module } from '@nestjs/common';
 import { IamModule as LibIamModule } from 'src/lib/bounded-contexts/iam/authentication/iam.module';
 import { MongoModule } from '@bitloops/bl-boilerplate-infra-mongo';
 import { PostgresModule } from '@bitloops/bl-boilerplate-infra-postgres';
-import { UserWriteRepoPortToken } from '@src/lib/bounded-contexts/iam/authentication/ports/user-write.repo-port';
 import { PubSubCommandHandlers } from '@src/lib/bounded-contexts/iam/authentication/application/command-handlers';
 import {
   JetstreamModule,
@@ -14,6 +13,7 @@ import { StreamingDomainEventHandlers } from '@src/lib/bounded-contexts/iam/auth
 import {
   StreamingDomainEventBusToken,
   StreamingIntegrationEventBusToken,
+  UserWriteRepoPortToken,
 } from '@src/lib/bounded-contexts/iam/authentication/constants';
 import { UserWritePostgresRepository } from './repository/user-write.pg.repository';
