+++ backend/src/bounded-contexts/iam/iam/repository/user-write.pg.repository.ts
@@ -9,9 +9,9 @@ import {
 import { Injectable, Inject } from '@nestjs/common';
 import * as jwtwebtoken from 'jsonwebtoken';
 import { UserWriteRepoPort } from '@src/lib/bounded-contexts/iam/authentication/ports/user-write.repo-port';
-import { UserEntity } from '@src/lib/bounded-contexts/iam/authentication/domain/UserEntity';
+import { UserEntity } from '@src/lib/bounded-contexts/iam/authentication/domain/user.entity';
 import { constants } from '@bitloops/bl-boilerplate-infra-postgres';
-import { EmailVO } from '@src/lib/bounded-contexts/iam/authentication/domain/EmailVO';
+import { EmailVO } from '@src/lib/bounded-contexts/iam/authentication/domain/email.value-object';
 import { ConfigService } from '@nestjs/config';
 import { AuthEnvironmentVariables } from '@src/config/auth.configuration';
 import { StreamingDomainEventBusToken } from '@src/lib/bounded-contexts/iam/authentication/constants';
