+++ backend/src/bounded-contexts/marketing/marketing/marketing.module.ts
@@ -2,12 +2,12 @@ import { Module } from '@nestjs/common';
 
 import { MarketingModule as LibMarketingModule } from 'src/lib/bounded-contexts/marketing/marketing/marketing.module';
 import { UserWriteRepository } from './repository/user-write.repository';
-import { UserWriteRepoPortToken } from '@src/lib/bounded-contexts/marketing/marketing/ports/user-write.repo-port';
-import { NotificationTemplateReadRepoPortToken } from '@src/lib/bounded-contexts/marketing/marketing/ports/notification-template-read.repo-port';
 import { NotificationTemplateReadRepository } from './repository/notification-template.repository';
 import {
   EmailServicePortToken,
+  NotificationTemplateReadRepoPortToken,
   StreamingCommandBusToken,
+  UserWriteRepoPortToken,
 } from '@src/lib/bounded-contexts/marketing/marketing/constants';
 import { MockEmailService } from './service';
 import { MongoModule } from '@bitloops/bl-boilerplate-infra-mongo';
