+++ backend/src/lib/bounded-contexts/iam/authentication/domain/email.value-object.ts
