+++ backend/src/lib/bounded-contexts/iam/authentication/domain/rules/index.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-namespace */
-import { ValidEmailRule as ValidEmail } from './ValidEmailRule';
+import { ValidEmailRule as ValidEmail } from './valid-email.rule';
 
 export namespace Rules {
   export class ValidEmailRule extends ValidEmail {}
