+++ backend/src/lib/bounded-contexts/iam/authentication/ports/user-write.repo-port.ts
@@ -1,6 +1,6 @@
 import { Application, Domain, Either } from '@bitloops/bl-boilerplate-core';
-import { EmailVO } from '../domain/EmailVO.js';
-import { UserEntity } from '../domain/UserEntity.js';
+import { EmailVO } from '../domain/email.value-object.js';
+import { UserEntity } from '../domain/user.entity.js';
 
 export interface UserWriteRepoPort
   extends Application.Repo.ICRUDWritePort<UserEntity, Domain.UUIDv4> {
@@ -8,5 +8,3 @@ export interface UserWriteRepoPort
     email: EmailVO,
   ): Promise<Either<UserEntity | null, Application.Repo.Errors.Unexpected>>;
 }
-
-export const UserWriteRepoPortToken = Symbol('UserWriteRepoPort');
