+++ backend/src/lib/bounded-contexts/marketing/marketing/constants.ts
@@ -1,3 +1,9 @@
 export const EmailServicePortToken = Symbol('EmailServicePort');
 
 export const StreamingCommandBusToken = Symbol('StreamingCommandBusToken');
+
+export const UserWriteRepoPortToken = Symbol('UserWriteRepoPort');
+
+export const NotificationTemplateReadRepoPortToken = Symbol(
+  'NotificationTemplateReadRepoPort',
+);
