+++ backend/src/lib/bounded-contexts/marketing/marketing/domain/email.value-object.ts
