+++ backend/src/lib/bounded-contexts/marketing/marketing/domain/notification-template.read-model.ts
