+++ backend/src/lib/bounded-contexts/marketing/marketing/domain/services/marketing-notification.service.ts
@@ -1,6 +1,6 @@
 import { Application, Either, ok, fail } from '@bitloops/bl-boilerplate-core';
 import { NotificationTemplateReadRepoPort } from '../../ports/notification-template-read.repo-port';
-import { NotificationTemplateReadModel } from '../read-models/notification-template.read-model';
+import { NotificationTemplateReadModel } from '../notification-template.read-model';
 import { UserEntity } from '../user.entity';
 
 export class MarketingNotificationService {
