+++ backend/src/lib/bounded-contexts/marketing/marketing/ports/user-write.repo-port.ts
@@ -5,5 +5,3 @@ export type UserWriteRepoPort = Application.Repo.ICRUDWritePort<
   UserEntity,
   Domain.UUIDv4
 >;
-
-export const UserWriteRepoPortToken = Symbol('UserWriteRepoPort');
