+++ backend/src/lib/bounded-contexts/marketing/marketing/tests/__tests__/todo-completions-incremented/notification-template-read-repo.mock.ts
@@ -5,7 +5,7 @@ import {
   fail,
   asyncLocalStorage,
 } from '@bitloops/bl-boilerplate-core';
-import { NotificationTemplateReadModel } from '../../../domain/read-models/notification-template.read-model';
+import { NotificationTemplateReadModel } from '../../../domain/notification-template.read-model';
 import { NotificationTemplateReadRepoPort } from '../../../ports/notification-template-read.repo-port';
 import {
   SUCCESS_CASE,
