+++ backend/src/lib/bounded-contexts/marketing/marketing/tests/builders/user-props.builder.ts
@@ -1,7 +1,7 @@
 import { Domain } from '@bitloops/bl-boilerplate-core';
 import { UserProps } from '@src/lib/bounded-contexts/marketing/marketing/domain/user.entity';
-import { CompletedTodosVO } from '@src/lib/bounded-contexts/marketing/marketing/domain/completed-todos.vo';
-import { EmailVO } from '../../domain/email.vo';
+import { CompletedTodosVO } from '@src/lib/bounded-contexts/marketing/marketing/domain/completed-todos.value-object';
+import { EmailVO } from '../../domain/email.value-object';
 
 export class UserPropsBuilder {
   private completedTodos: number;
