+++ backend/src/lib/bounded-contexts/todo/todo/constants.ts
@@ -5,3 +5,7 @@ export const PubSubIntegrationEventBusToken = Symbol(
   'PubSubIntegrationEventBusToken',
 );
 export const StreamingDomainEventBusToken = Symbol('StreamingDomainEventBus');
+
+export const TodoWriteRepoPortToken = Symbol('TodoWriteRepoPort');
+
+export const TodoReadRepoPortToken = Symbol('TodoReadRepoPort');
