+++ backend/src/lib/bounded-contexts/todo/todo/ports/todo-write.repo-port.ts
@@ -1,5 +1,5 @@
 import { Application, Domain } from '@bitloops/bl-boilerplate-core';
-import { TodoEntity } from '../domain/TodoEntity.js';
+import { TodoEntity } from '../domain/todo.entity.js';
 
 export type TodoWriteRepoPort = Application.Repo.ICRUDWritePort<
   TodoEntity,
@@ -11,5 +11,3 @@ export type TodoWriteRepoPort = Application.Repo.ICRUDWritePort<
 //   findAll(): Promise<TodoEntity[]>;
 //   save(hero: TodoEntity): Promise<void>;
 // }
-
-export const TodoWriteRepoPortToken = Symbol('TodoWriteRepoPort');
