+++ backend/src/lib/bounded-contexts/todo/todo/tests/__tests__/add-todo/add-todo-write-repo.mock.ts
@@ -5,7 +5,7 @@ import {
   fail,
   ok,
 } from '@bitloops/bl-boilerplate-core';
-import { TodoEntity } from '@src/lib/bounded-contexts/todo/todo/domain/TodoEntity';
+import { TodoEntity } from '@src/lib/bounded-contexts/todo/todo/domain/todo.entity';
 import { TodoWriteRepoPort } from '@src/lib/bounded-contexts/todo/todo/ports/todo-write.repo-port';
 import { ADD_TODO_REPO_ERROR_CASE } from './add-todo.mock';
