+++ backend/src/lib/bounded-contexts/todo/todo/tests/__tests__/complete-todo/complete-todo-write-repo.mock.ts
@@ -5,7 +5,7 @@ import {
   ok,
   fail,
 } from '@bitloops/bl-boilerplate-core';
-import { TodoEntity } from '@src/lib/bounded-contexts/todo/todo/domain/TodoEntity';
+import { TodoEntity } from '@src/lib/bounded-contexts/todo/todo/domain/todo.entity';
 import { TodoWriteRepoPort } from '@src/lib/bounded-contexts/todo/todo/ports/todo-write.repo-port';
 import { DomainErrors } from '@src/lib/bounded-contexts/todo/todo/domain/errors';
 import {
